Require a 'title' element to be passed to run_cpanm().
For now, this element will simply be used to compose the names of the gzipped
build.log, files in analysis/, storage/, etc.

Ideally, we would like to track the release-to-release progress of different
sets of CPAN or other modules, in which case 'title' would mean 'title of this
tracking project'.  But to do that we'd have to rework 'cpanm' to not put all
of its logs in directories simply distinguised by timestamps.

Todo: A given monthly dev release may require more than one run of cpanm,
particularly when setting up a baseline.  To do that, we have to calculate a
'run' number for each gzipped log file.  (I've done that locally in shell
scripts, but not yet for this git repo.)

Aside from the Todo, this essentially completes the translation of
perform_cpanm_cpan_river to perl.