Remove certain attributes from object.  Pass application_dir rather than results_dir.
For consistency between the two constructors, we'll pass 'application_dir'
rather than 'results_dir' in new_from_existing_perl_cpanm().  In the latter
we'll check that the value of application_dir exists, then create results_dir
and testing_dir if they don't already exist.  This will facilitate comparison of the
contents of the objects created by the different constructors.

For the same reason, remove from the object 4 attributes -- host, hostdir,
compression, tarball -- that are only used within the scope of
perform_perl_download().