Eliminate concept of 'run'.
Originally, I needed the concept of 'run' because I needed multiple
invocations of cpanm for a given monthly release to get meaningful results.
Why? IIRC, because I had not yet fully addressed problems like missing
external libraries, badly behaved distros etc.

The concept of 'run' first appears when, inside of run_cpanm(), we gzip the
log files.  It cascades through later methods.  If we eliminate it we simplify
the source code and simplify the directory structure we have to work with.