Assign cpanm_dir within constructor.
Within the alternate constructor, new_from_existing_perl_cpanm(), test for the
existence of the $release_dir/.cpanm directory, which should have been created
during a previous run of the caller.  Assuming it exists, asssign it to the
'cpanm_dir' KVP within the object so that it can be located via
'get_cpanm_dir()'.  Add tests for this.